requirements.txt
# --- Core Backend ---
fastapi>=0.129.0
uvicorn>=0.40.0
python-multipart>=0.0.22
pydantic>=2.12.5
pydantic-settings>=2.13.0
python-dotenv>=1.2.1
psycopg2-binary>=2.9.11

# --- Telegram Bot ---
pyTelegramBotAPI>=4.31.0
requests>=2.32.5

# --- PDF Processing ---
PyMuPDF>=1.27.1
pypdf>=5.1.0

# --- AI & Vector Search ---
chromadb>=1.5.0
sentence-transformers>=5.2.2
huggingface_hub>=0.36.2

# --- Utilities ---
pandas>=2.3.3
numpy>=2.4.2
httpx>=0.28.1
aiofiles>=23.2.1
pillow>=12.0.0
pdf2image>=1.17.0